Support Multiple Artists per Song
Most songs have more than one artist in them, so having native support
for these will be nice. On the client side, a song can be creditted to
many artists with links to more of their work, instead of their name
being in the title.

Some of the packages were reagranged. Server-specific files were put
under the server package.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>