Check if the Song exists before adding it to the Database
There are some checks to see if a songs already exists in the database.
If a song with the same ID is there that means the same file is trying
to be uploaded. If it has the same title & artists, then it's a
different version (or quality or something) of the same song.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>