Move IDs from Song to SongInfo and respond with Infos
If we want the details of the song with the ID, we have to send the
song bytes with it. This is a problem when we want to do things like
searching, which requires the details and the id, but not the song
bytes.

A solution to this problem is simply moving the ID into the information
of the song, and keeping the Song structure to the bytes with the info.
This will let us send SongInfo packets, which will be small compared to
the entire song, without the sound bytes.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>